test llama3

+++ runs2.py
@@ -82,7 +82,7 @@ def probe_healthy_gpus(min_free_gb: float = 4.0, compute_test: bool = True) -> l
 DATASET_PATH = "data/AdvBench_100.csv"
 N_GOALS = None
 RANDOM_SEED = 42
-PRECISION = "int8"
+PRECISION = "fp16"
 
 CHECKPOINT_PATH = f"results/checkpoint_{PRECISION}.json"
